pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webhook-delivery-service"
version = "0.1.0"
description = "Production-grade webhook delivery infrastructure. Multi-tenant event ingestion, reliable HTTP delivery with retries, circuit breaker, HMAC signing, rate limiting, and semantic log search via pgvector."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "structlog>=25.4.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.4",
    "pre-commit>=4.2.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["alembic/versions/*"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "B",
    "S",
    "C4",
    "UP",
    "SIM",
    "PTH",
    "RUF",
]
ignore = [
    "E501",
    "S101",
    "S104",
    "B008",
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.isort]
known-first-party = ["webhook_delivery_service"]
split-on-trailing-comma = true
force-wrap-aliases = true
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
strict = true
pretty = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
